Move Android sources to `android/` subdirectory and update includes
