+++ packages/react-native-reanimated/android/src/main/cpp/reanimated/android/AndroidUIScheduler.cpp
@@ -1,4 +1,4 @@
-#include <reanimated/AndroidUIScheduler.h>
+#include <reanimated/android/AndroidUIScheduler.h>
 
 #include <android/log.h>
 #include <fbjni/fbjni.h>
