+++ packages/react-native-reanimated/android/src/main/cpp/reanimated/android/LayoutAnimations.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <reanimated/JNIHelper.h>
+#include <reanimated/android/JNIHelper.h>
 
 #include <fbjni/fbjni.h>
 #include <jsi/jsi.h>
